+++ packages/backend/src/accounting/psql/ledger-account/index.test.ts
@@ -27,6 +27,7 @@ describe('Ledger Account', (): void => {
     deps = initIocContainer({ ...Config, useTigerBeetle: false })
     appContainer = await createTestApp(deps)
     serviceDeps = {
+      config: await deps.use('config'),
       logger: await deps.use('logger'),
       knex: await deps.use('knex'),
       telemetry: await deps.use('telemetry')
