+++ packages/backend/src/accounting/psql/ledger-transfer/index.test.ts
@@ -36,6 +36,7 @@ describe('Ledger Transfer', (): void => {
     deps = initIocContainer({ ...Config, useTigerBeetle: false })
     appContainer = await createTestApp(deps)
     serviceDeps = {
+      config: await deps.use('config'),
       logger: await deps.use('logger'),
       knex: await deps.use('knex'),
       telemetry: await deps.use('telemetry')
