+++ packages/backend/src/open_payments/wallet_address/model.ts
@@ -4,7 +4,7 @@ import { LiquidityAccount, OnCreditOptions } from '../../accounting/service'
 import { ConnectorAccount } from '../../payment-method/ilp/connector/core/rafiki'
 import { Asset } from '../../asset/model'
 import { BaseModel, Pagination, SortOrder } from '../../shared/baseModel'
-import { WebhookEvent } from '../../webhook/model'
+import { WebhookEvent } from '../../webhook/event/model'
 import { WalletAddressKey } from '../../open_payments/wallet_address/key/model'
 import { AmountJSON } from '../amount'
 import { WalletAddressAdditionalProperty } from './additional_property/model'
