+++ packages/backend/src/tests/webhook.ts
@@ -2,7 +2,7 @@ import { faker } from '@faker-js/faker'
 import { v4 as uuid } from 'uuid'
 import { IocContract } from '@adonisjs/fold'
 import { AppServices } from '../app'
-import { WebhookEvent } from '../webhook/model'
+import { WebhookEvent } from '../webhook/event/model'
 import { sample } from 'lodash'
 import { EventPayload } from '../webhook/service'
 import { createAsset } from './asset'
