+++ packages/backend/src/webhook/service.ts
@@ -2,8 +2,8 @@ import axios, { isAxiosError } from 'axios'
 import { createHmac } from 'crypto'
 import { canonicalize } from 'json-canonicalize'
 
-import { isWebhookWithEvent, Webhook, WebhookWithEvent } from './hook/model'
-import { WebhookEvent } from './model'
+import { isWebhookWithEvent, Webhook, WebhookWithEvent } from './model'
+import { WebhookEvent } from './event/model'
 import { IAppConfig } from '../config/app'
 import { BaseService } from '../shared/baseService'
 import { Pagination, SortOrder } from '../shared/baseModel'
